Final/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
python-dotenv==1.0.1
requests==2.31.0
pydantic==2.5.0
boto3==1.34.0
botocore==1.34.0
